test: move a module to its own file
